+++ CHANGES.md
@@ -9,7 +9,17 @@
 - FIX
     - バグ修正
 
-## develop
+## 2020.2.0
+- [UPDATE] sora-js-sdk のバージョンを 2020.3.0 に更新
+    - Safari の Simulcast 対応
+    - @yuitowest
+- [ADD] Form から metadata を指定できる機能を追加
+    - @yuitowest
+- [UPDATE] PeerConnection getStats のレポートを componet 内から Redux state へ移動
+    - Download Report から出力される json に最後に取得した PeerConnection getStats のレポートを記載
+    - @yuitowest
+- [ADD] Debug カラムに push message を表示するタブを追加
+    - @yuitowest
 
 ## 2020.1.0